requirements.txt
Django==5.2.6
gunicorn==23.0.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.1.0
scikit-image==0.25.2
joblib==1.4.2
python-dotenv==1.0.1
google-generativeai==0.8.4